add trace autoconfiguration feature
