+++ google-cloud-graalvm-samples/spring-cloud-gcp-trace-sample/src/main/resources/META-INF/native-image/com.example/proxy-config.json
@@ -1,0 +1,3 @@
+[
+    ["com.example.WorkService", "org.springframework.aop.SpringProxy", "org.springframework.aop.framework.Advised", "org.springframework.core.DecoratingProxy"]
+]
